@@ -3,32 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.1.1",
-    "@reduxjs/toolkit": "^1.5.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/luxon": "^1.26.2",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router-dom": "^5.1.7",
+    "@reduxjs/toolkit": "^1.9.7",
     "antd": "^4.13.1",
     "axios": "^0.21.1",
-    "connected-react-router": "^6.9.1",
-    "craco-less": "^1.17.1",
+    "craco-less": "^3.0.1",
+    "history": "^5.3.0",
     "luxon": "^1.26.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-intl": "^5.13.2",
-    "react-redux": "^7.2.2",
-    "react-responsive": "^8.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.0.5",
-    "typescript": "^4.1.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-intl": "^6.5.1",
+    "react-redux": "^8.1.3",
+    "react-responsive": "^9.0.2",
+    "react-router-dom": "^6.18.0",
+    "redux": "^4.2.1",
+    "redux-first-history": "^5.1.1",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -38,10 +26,17 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": ["react-app", "react-app/jest"]
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
-    "production": [">0.2%", "not dead", "not op_mini all"],
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
     "development": [
       "last 1 chrome version",
       "last 1 firefox version",
@@ -49,6 +44,20 @@
     ]
   },
   "devDependencies": {
-    "@types/react-responsive": "^8.0.2"
+    "@craco/craco": "^7.1.0",
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^12.1.10",
+    "@types/jest": "^26.0.15",
+    "@types/luxon": "^1.26.2",
+    "@types/node": "^12.0.0",
+    "@types/react": "^18.2.35",
+    "@types/react-dom": "^18.2.14",
+    "@types/react-responsive": "^8.0.2",
+    "react-scripts": "5.0.1",
+    "typescript": "^5.2.2"
+  },
+  "resolutions": {
+    "**/@types/react": "^18.2.35"
   }
 }
