@@ -1,35 +1,35 @@
-import './App.less';
-import { Button, Layout, Result, Spin } from 'antd';
-import { Route, Switch } from 'react-router-dom';
-import Login from './Login';
-import { useDispatch, useSelector } from 'react-redux';
-import React from 'react';
-import { getAccount } from './Login/state';
-import Home from './Home';
+import "./App.less";
+import { Button, Layout, Result, Spin } from "antd";
+import { Route, Routes } from "react-router-dom";
+import { HistoryRouter as Router } from "redux-first-history/rr6";
+import Login from "./Login";
+import { useAppDispatch, useAppSelector } from "./hooks";
+import { history } from "./store";
+import React from "react";
+import { getAccount } from "./Login/state";
+import Home from "./Home";
 import Consultations, {
   Consultation,
   NewConsultationModal,
-} from './Consultations';
-import Pets, { NewPetModal, Pet } from './Pets';
-import Owners, { NewOwnerModal, Owner } from './Owners';
-import { ConnectedRouter } from 'connected-react-router';
-import { history } from './redux/store';
-import { NewActModal, UpdateActModal } from './Acts';
-import Settings from './Settings';
-import Dashboard from './Dashboard';
-import Navigation from './Navigation';
+} from "./Consultations";
+import Pets, { NewPetModal, Pet } from "./Pets";
+import Owners, { NewOwnerModal, Owner } from "./Owners";
+import { NewActModal, UpdateActModal } from "./Acts";
+import Settings from "./Settings";
+import Dashboard from "./Dashboard";
+import Navigation from "./Navigation";
 
 const App = () => {
-  const account = useSelector((s) => s.login.account);
-  const dispatch = useDispatch();
+  const account = useAppSelector((s) => s.login.account);
+  const dispatch = useAppDispatch();
   React.useEffect(() => {
-    if (account === 'NotRequested') {
+    if (account === "NotRequested") {
       dispatch(getAccount());
     }
   }, [account, dispatch]);
 
   return (
-    <ConnectedRouter history={history}>
+    <Router history={history}>
       <NewConsultationModal />
       <NewPetModal />
       <NewOwnerModal />
@@ -38,25 +38,21 @@ const App = () => {
       <Layout>
         <Navigation />
         <Layout.Content>
-          {typeof account === 'object' ? (
-            <Switch>
-              <Route
-                path="/consultations/:key"
-                exact
-                component={Consultation}
-              />
-              <Route path="/consultations" exact component={Consultations} />
-              <Route path="/owners/:key" exact component={Owner} />
-              <Route path="/owners" exact component={Owners} />
-              <Route path="/pets/:key" exact component={Pet} />
-              <Route path="/pets" exact component={Pets} />
-              <Route path="/dashboard" exact component={Dashboard} />
-              <Route path="/settings" exact component={Settings} />
-              <Route path="/" exact component={Home} />
-            </Switch>
-          ) : account === 'Outsider' ? (
+          {typeof account === "object" ? (
+            <Routes>
+              <Route path="/consultations/:key" element={<Consultation />} />
+              <Route path="/consultations" element={<Consultations />} />
+              <Route path="/owners/:key" element={<Owner />} />
+              <Route path="/owners" element={<Owners />} />
+              <Route path="/pets/:key" element={<Pet />} />
+              <Route path="/pets" element={<Pets />} />
+              <Route path="/dashboard" element={<Dashboard />} />
+              <Route path="/settings" element={<Settings />} />
+              <Route path="/" element={<Home />} />
+            </Routes>
+          ) : account === "Outsider" ? (
             <Login />
-          ) : account === 'NetworkError' ? (
+          ) : account === "NetworkError" ? (
             <Result
               title="Erreur"
               subTitle="Une erreur est survenue, veuillez contacter Wael"
@@ -71,7 +67,7 @@ const App = () => {
           )}
         </Layout.Content>
       </Layout>
-    </ConnectedRouter>
+    </Router>
   );
 };
 
