@@ -4,7 +4,7 @@ module.exports = {
   devServer: {
     proxy: {
       "/api": {
-        target: "http://localhost:1234",
+        target: "http://127.0.0.1:1234",
         secure: false,
         changeOrigin: true,
       },
