@@ -1,50 +1,55 @@
-import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { DefaultRootState } from 'react-redux';
-import { createAct, updateAct } from '../api';
-import { logout } from '../Login/state';
-import { searchActsAction } from '../Settings';
-import { Act, Keyed, Loadable } from '../types';
+import { createAsyncThunk, createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { createAct, updateAct } from "../api";
+import { searchActsAction, LocalState as SettingsState } from "../Settings";
+import { Act, Keyed, Loadable } from "../types";
 
-export const createActAction = createAsyncThunk<
-  Keyed<Act>,
-  Act,
-  { state: DefaultRootState }
->('acts/create', async (p: Act, thunkAPI) => {
-  const res = await createAct(p);
-  thunkAPI.dispatch(searchActsAction(thunkAPI.getState().settings.search));
-  return res.data;
-});
+const initialState = {
+  createModalOpen: false,
+  createState: "NotRequested" as Loadable<"Ok">,
+  selectedAct: undefined as Keyed<Act> | undefined,
+  updateState: "NotRequested" as Loadable<"Ok">,
+};
+
+export type LocalState = typeof initialState;
+type ThunkAPI = {
+  state: {
+    settings: SettingsState;
+    acts: LocalState;
+  };
+};
+
+export const createActAction = createAsyncThunk<Keyed<Act>, Act, ThunkAPI>(
+  "acts/create",
+  async (p: Act, thunkAPI) => {
+    const res = await createAct(p);
+    thunkAPI.dispatch(searchActsAction(thunkAPI.getState().settings.search));
+    return res.data;
+  },
+);
 
 export const updateActAction = createAsyncThunk<
   Keyed<Act>,
   Keyed<Act>,
-  { state: DefaultRootState }
->('acts/update', async (p: Keyed<Act>, thunkAPI) => {
+  ThunkAPI
+>("acts/update", async (p: Keyed<Act>, thunkAPI) => {
   const res = await updateAct(p);
   thunkAPI.dispatch(searchActsAction(thunkAPI.getState().settings.search));
   return res.data;
 });
 
-const initialState = {
-  createModalOpen: false,
-  createState: 'NotRequested' as Loadable<'Ok'>,
-  selectedAct: undefined as Keyed<Act> | undefined,
-  updateState: 'NotRequested' as Loadable<'Ok'>,
-};
-
 const actsSlice = createSlice({
-  name: 'acts',
+  name: "acts",
   initialState,
   reducers: {
     openNewActModal: (s) => {
-      s.createState = 'NotRequested';
+      s.createState = "NotRequested";
       s.createModalOpen = true;
     },
     closeActModal: (s) => {
       s.createModalOpen = false;
     },
     openUpdateActModal: (s, a: PayloadAction<Keyed<Act>>) => {
-      s.updateState = 'NotRequested';
+      s.updateState = "NotRequested";
       s.selectedAct = a.payload;
     },
     closeUpdateActModal: (s) => {
@@ -54,26 +59,26 @@ const actsSlice = createSlice({
   extraReducers: (b) =>
     b
       .addCase(createActAction.pending, (s) => {
-        s.createState = 'Loading';
+        s.createState = "Loading";
       })
       .addCase(createActAction.fulfilled, (s) => {
-        s.createState = 'Ok';
+        s.createState = "Ok";
         s.createModalOpen = false;
       })
       .addCase(createActAction.rejected, (s) => {
-        s.createState = 'NetworkError';
+        s.createState = "NetworkError";
       })
       .addCase(updateActAction.pending, (s) => {
-        s.updateState = 'Loading';
+        s.updateState = "Loading";
       })
       .addCase(updateActAction.fulfilled, (s) => {
-        s.updateState = 'Ok';
+        s.updateState = "Ok";
         s.selectedAct = undefined;
       })
       .addCase(updateActAction.rejected, (s) => {
-        s.updateState = 'NetworkError';
+        s.updateState = "NetworkError";
       })
-      .addCase('logout/pending', () => initialState),
+      .addCase("logout/pending", () => initialState),
 });
 
 export const {
