@@ -1,13 +1,19 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { DateTime } from 'luxon';
-import { searchConsultations, getConsultationStatistics } from '../api';
-import { logout } from '../Login/state';
-import {
-  ConsultationStatistics,
-  FullConsultation,
-  initMemo,
-  WithMemo,
-} from '../types';
+import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import { DateTime } from "luxon";
+import { searchConsultations, getConsultationStatistics } from "../api";
+import { ConsultationStatistics, FullConsultation, initMemo } from "../types";
+
+const initialState = {
+  consultations: initMemo<FullConsultation[]>([]),
+  statistics: initMemo<ConsultationStatistics>({
+    remainingConsultations: 0,
+    doneConsultations: 0,
+    earnings: 0,
+  }),
+};
+
+export type LocalState = typeof initialState;
+type ThunkAPI = { state: { home: LocalState } };
 
 export const getHomeData = createAsyncThunk<
   {
@@ -15,15 +21,15 @@ export const getHomeData = createAsyncThunk<
     consultations: FullConsultation[];
   },
   void,
-  { state: { home: { consultations: WithMemo<any> } } }
+  ThunkAPI
 >(
-  'getHomeData',
+  "getHomeData",
   async () => {
     const now = DateTime.local();
     const d = await searchConsultations({
       after: now.set({ hour: 0, minute: 0, second: 0, millisecond: 0 }),
       before: now.set({ hour: 24, minute: 0, second: 0, millisecond: 0 }),
-      status: 'pending',
+      status: "pending",
     });
     const statistics = await getConsultationStatistics();
     return { statistics, consultations: d.rows };
@@ -31,29 +37,20 @@ export const getHomeData = createAsyncThunk<
   {
     condition: (_, { getState }) => {
       const { home } = getState();
-      return home.consultations.value !== 'Loading';
+      return home.consultations.value !== "Loading";
     },
-  }
+  },
 );
 
-  const initialState= {
-    consultations: initMemo<FullConsultation[]>([]),
-    statistics: initMemo<ConsultationStatistics>({
-      remainingConsultations: 0,
-      doneConsultations: 0,
-      earnings: 0,
-    }),
-  }
-
 const homeSlice = createSlice({
-  name: 'home',
-  initialState ,
+  name: "home",
+  initialState,
   reducers: {},
   extraReducers: (b) =>
     b
       .addCase(getHomeData.pending, (s) => {
-        s.consultations.value = 'Loading';
-        s.statistics.value = 'Loading';
+        s.consultations.value = "Loading";
+        s.statistics.value = "Loading";
       })
       .addCase(getHomeData.fulfilled, (s, a) => {
         s.consultations.value = a.payload.consultations;
@@ -62,10 +59,10 @@ const homeSlice = createSlice({
         s.statistics.memo = a.payload.statistics;
       })
       .addCase(getHomeData.rejected, (s) => {
-        s.consultations.value = 'NetworkError';
-        s.statistics.value = 'NetworkError';
+        s.consultations.value = "NetworkError";
+        s.statistics.value = "NetworkError";
       })
-      .addCase('logout/pending', () => initialState),
+      .addCase("logout/pending", () => initialState),
 });
 
 export default homeSlice.reducer;
