@@ -1,7 +1,5 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { DefaultRootState } from 'react-redux';
-import { createPet, getPet, searchPets, updatePet } from '../api';
-import { logout } from "../Login/state";
+import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import { createPet, getPet, searchPets, updatePet } from "../api";
 import {
   Pet,
   Loadable,
@@ -12,22 +10,40 @@ import {
   initMemo,
   WithMemo,
   FullPet,
-} from '../types';
+} from "../types";
+
+const initialState = {
+  createModalOpen: false,
+  createState: "NotRequested" as Loadable<Owned<Keyed<Pet>>>,
+  searchRequest: { page: 0, itemsPerPage: 10 } as SearchRequest,
+  pets: initMemo({ rows: [], total: 0 }) as WithMemo<
+    SearchResult<Keyed<Owned<Pet>>>
+  >,
+  current: "NotRequested" as Loadable<FullPet>,
+  updateState: "NotRequested" as Loadable<"Ok">,
+};
+
+export type LocalState = typeof initialState;
+type ThunkAPI = {
+  state: {
+    pets: LocalState;
+  };
+};
 
 export const createPetAction = createAsyncThunk(
-  'pets/createPet',
+  "pets/createPet",
   async (p: Owned<Pet>, thunkAPI) => {
     const res = await createPet(p);
     thunkAPI.dispatch(getPets({}) as any);
     return res;
-  }
+  },
 );
 
 export const getPets = createAsyncThunk<
   SearchResult<Keyed<Owned<Pet>>>,
   SearchRequest,
-  { state: DefaultRootState }
->('pets/getPets', (p: SearchRequest, thunkAPI) => {
+  ThunkAPI
+>("pets/getPets", (p: SearchRequest, thunkAPI) => {
   const req: SearchRequest = {
     ...thunkAPI.getState().pets.searchRequest,
     ...p,
@@ -37,29 +53,18 @@ export const getPets = createAsyncThunk<
 });
 
 export const selectPetAction = createAsyncThunk(
-  'pets/selectPet',
-  (pKey: number) => getPet(pKey)
+  "pets/selectPet",
+  (pKey: number) => getPet(pKey),
 );
 
-export const updatePetAction = createAsyncThunk('pet/updatePet', updatePet);
-
-const initialState = {
-  createModalOpen: false,
-  createState: 'NotRequested' as Loadable<Owned<Keyed<Pet>>>,
-  searchRequest: { page: 0, itemsPerPage: 10 } as SearchRequest,
-  pets: initMemo({ rows: [], total: 0 }) as WithMemo<
-    SearchResult<Keyed<Owned<Pet>>>
-  >,
-  current: 'NotRequested' as Loadable<FullPet>,
-  updateState: 'NotRequested' as Loadable<'Ok'>,
-};
+export const updatePetAction = createAsyncThunk("pet/updatePet", updatePet);
 
 const petsSlice = createSlice({
-  name: 'pets',
+  name: "pets",
   initialState,
   reducers: {
     openNewPetModal: (s) => {
-      s.createState = 'NotRequested';
+      s.createState = "NotRequested";
       s.createModalOpen = true;
     },
     closePetModal: (s) => {
@@ -69,45 +74,45 @@ const petsSlice = createSlice({
   extraReducers: (b) =>
     b
       .addCase(createPetAction.pending, (s) => {
-        s.createState = 'Loading';
+        s.createState = "Loading";
       })
       .addCase(createPetAction.fulfilled, (s, a) => {
         s.createState = a.payload;
         s.createModalOpen = false;
       })
       .addCase(createPetAction.rejected, (s) => {
-        s.createState = 'NetworkError';
+        s.createState = "NetworkError";
       })
       .addCase(getPets.pending, (s, a) => {
         s.searchRequest = { ...s.searchRequest, ...a.meta.arg };
-        s.pets.value = 'Loading';
+        s.pets.value = "Loading";
       })
       .addCase(getPets.fulfilled, (s, a) => {
         s.pets.value = a.payload;
         s.pets.memo = a.payload;
       })
       .addCase(getPets.rejected, (s) => {
-        s.pets.value = 'NetworkError';
+        s.pets.value = "NetworkError";
       })
       .addCase(selectPetAction.pending, (s) => {
-        s.current = 'Loading';
+        s.current = "Loading";
       })
       .addCase(selectPetAction.fulfilled, (s, a) => {
         s.current = a.payload;
       })
       .addCase(selectPetAction.rejected, (s) => {
-        s.current = 'NetworkError';
+        s.current = "NetworkError";
       })
       .addCase(updatePetAction.pending, (s) => {
-        s.updateState = 'Loading';
+        s.updateState = "Loading";
       })
       .addCase(updatePetAction.fulfilled, (s) => {
-        s.updateState = 'Ok';
+        s.updateState = "Ok";
       })
       .addCase(updatePetAction.rejected, (s) => {
-        s.updateState = 'NetworkError';
+        s.updateState = "NetworkError";
       })
-      .addCase('logout/pending', () => initialState),
+      .addCase("logout/pending", () => initialState),
 });
 
 export const { openNewPetModal, closePetModal } = petsSlice.actions;
