@@ -4,7 +4,7 @@ import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { Provider } from "react-redux";
-import store from "./redux/store";
+import store from "./store";
 import { IntlProvider } from "react-intl";
 
 ReactDOM.render(
@@ -15,7 +15,7 @@ ReactDOM.render(
       </React.StrictMode>
     </IntlProvider>
   </Provider>,
-  document.getElementById("root")
+  document.getElementById("root"),
 );
 
 // If you want to start measuring performance in your app, pass a function
